Cleaned up ConstraintStatistics and MeansVars
Made it so we aren't feeding unnecessary enums into these functions. I
can do this after realising that the m_state_x and m_state_y arguments
in the BoxLoop define the scope of the class, so starting off in
m_state_new allows the class to read evolution variables, ending in
m_state_diagnostic allows the class to store values in diagnostic
variables, and vice versa.

Note that the MeansVars class is not quite simple, could be a good use
case for future students who just want to do simple manipulations at
every point in the box.