Incorporated changes from BD code, commit c9392d01bbd43d04f132c7c76a5156f1cb295a4d and up. Briefly:
1. Made just two random draws per point per field, one for modulus and one for argument.
This turns the draw at each point from a Rayleigh for the real/im components into a
decomposition of the modulus/argument form of the field onto the Cartesian form.
2. Made the norm 2\pi/L instead of 1/L, and changed A from 1e-4 to 1e-6 to compensate.
3. Added data_path as an argument, so the stdev file for the polarisation fields gets put
into the data folder on each run.
4. Took out the sqrt from the find_rayleigh_factor function, and put the result of that
fn in the sqrt draw for the modulus. This should be the right way to do things, based on my
notebook notes from today.