Fixed random angle bug and added timing
1. Fixed a bug with how the random angle was applied to the initial
random field. Basically, foiled the complex number a + bi over the
angular factor cos(theta) + isin(theta) outside of the find_rayleigh_factor
function. This produces a match in the evolution of the binned PS AND in
individual modes in the Fourier box.
2. Did a tiny bit of cleaning on the RF class.
3. Added print of run duration, based off of the BinaryBH example, to the Main
file. Haven't tested this in the run yet, but it does compile.
4. Set up params for longer run (out to efold 8)