(!!) Last commit before paper 1 submission
This was the last commit on this branch (single-mode-test).
In this commit I included the stochastic phase for both single-mode
tests which differ on wheather the MS phase is used.

- Redid this test so that only one mode is initialised (as before, at
i,j,k = 4,0,0 currently), but where turning use_rand on and off actually
changes whether or not the mode function phase is included.
- Off means just the stochastic phase is used, and on means both are used.
- For now using the Rayleigh draw is turned off, as this just introduces
an overall factor in the phase match between the numerical results
and MS solution.